+++ admin/tabs/activity/tracking/ActivityGrid.js
@@ -12,7 +12,7 @@ import {panel} from '@xh/hoist/desktop/cmp/panel';
 import {dateInput, textInput} from '@xh/hoist/desktop/cmp/input';
 import {toolbarSep} from '@xh/hoist/desktop/cmp/toolbar';
 import {button, exportButton, refreshButton} from '@xh/hoist/desktop/cmp/button';
-import {storeCountLabel} from '@xh/hoist/desktop/cmp/store';
+import {storeCountLabel} from '@xh/hoist/cmp/store';
 import {Icon} from '@xh/hoist/icon';
 
 import {ActivityGridModel} from './ActivityGridModel';